Add files via upload
I do not own the copyright to the video. It is used for testing purposes. Higher quality videos work better. I also added tube jackal so that you may download your own videos. See my other repo for instructions. Please change the path of the SOURCE file to your own path so the demo may work. To experiment, use the notebook file. When you are satisfied with the results use the object_detection.py file.